Doing some refactoring to resubmit my Trees assignment since it wasn't formatted the way the TAs wanted.
