+++ DataStructuresAndSorts/Assignment5/Assignment5 - Trees.csproj
@@ -38,6 +38,7 @@
   <ItemGroup>
     <Compile Include="BinarySearchTree.cs" />
     <Compile Include="Caller.cs" />
+    <Compile Include="ExpressionParser.cs" />
     <Compile Include="Hero.cs" />
     <Compile Include="myComparers.cs" />
     <Compile Include="ExpressionTree.cs" />
